Extract ShellEscape module.
