+++ lib/rspec/core/rake_task.rb
@@ -1,6 +1,7 @@
 require 'rake'
 require 'rake/tasklib'
 require 'rspec/support/ruby_features'
+require 'rspec/core/shell_escape'
 
 module RSpec
   module Core
@@ -9,6 +10,7 @@ module RSpec
     # @see Rakefile
     class RakeTask < ::Rake::TaskLib
       include ::Rake::DSL if defined?(::Rake::DSL)
+      include RSpec::Core::ShellEscape
 
       # Default path to the RSpec executable.
       DEFAULT_RSPEC_PATH = File.expand_path('../../../../exe/rspec', __FILE__)
@@ -119,20 +121,6 @@ module RSpec
           # TODO: consider deprecating support for this and removing it in
           #   RSpec 4.
           FileList[pattern].sort.map { |file| escape file }
-        end
-      end
-
-      if RSpec::Support::OS.windows?
-        # :nocov:
-        def escape(shell_command)
-          "'#{shell_command.gsub("'", "\\\\'")}'"
-        end
-        # :nocov:
-      else
-        require 'shellwords'
-
-        def escape(shell_command)
-          shell_command.shellescape
         end
       end
